[No JIRA] Fix BpkImage onload when server-side rendering
When BpkImage is used with server-side rendering, there was a race
condition where if the underlying img element loaded before the React
bundle executed and attached the onload listener, the image would appear
to be loading forever.

Fix it by checking whether the image is already loaded when the
component mounts and triggering the callback if it is.