[No JIRA] Generate duplicate copy of icon font
We need to generate a second copy of the icon font with a different font
name to allow two copies of the font to be loaded in the same app. The
reason we need two copies is one is for Backpack React Native and the other is
for Backpack iOS. Crucially the name of the font as recorded in the
`name` table must be different to allow UIKit to load the font twice.